refactor: reorganize question bank and student management components into modular view and dialog structures

+++ app/sentinel-api/src/modules/question-collection/data/get-question-collections.ts
@@ -37,7 +37,7 @@ export async function getQuestionCollectionsData({
             'creator.last_name as creator_last_name',
             'updater.first_name as updater_first_name',
             'updater.last_name as updater_last_name',
-            sql<number>`count(qcq.question_bank_question_id)`.as('question_count'),
+            sql<number>`count(qcq.question_bank_question_id)::int`.as('question_count'),
         ])
         .groupBy([
             'qc.collection_id',
